+++ src/ipfs/get.js
@@ -1,0 +1,15 @@
+import fetch from 'cross-fetch';
+import { DEFAULT_IPFS_GATEWAY } from '../conf.js';
+
+const get = async (cid, { ipfsGateway = DEFAULT_IPFS_GATEWAY } = {}) => {
+  const multiaddr = `/ipfs/${cid.toString()}`;
+  const publicUrl = `${ipfsGateway}${multiaddr}`;
+  const res = await fetch(publicUrl);
+  if (!res.ok) {
+    throw Error(`Failed to load content from ${publicUrl}`);
+  }
+  const arrayBuffer = await res.arrayBuffer();
+  return new Uint8Array(arrayBuffer);
+};
+
+export default get;
