+++ src/ipfs/isCid.js
@@ -1,0 +1,12 @@
+import CID from 'cids';
+
+const isCid = (value) => {
+  try {
+    const cid = new CID(value);
+    return CID.isCID(cid);
+  } catch (e) {
+    return false;
+  }
+};
+
+export default isCid;
